requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
datasets==2.20.0
dill==0.3.8
dnspython==2.6.1
efficientnet_pytorch==0.7.1
einops==0.8.0
exceptiongroup==1.2.2
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.2
idna==3.7
intel-openmp==2021.4.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
linformer==0.2.3
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
preshed==3.0.9
protobuf==5.27.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12.1
tbb==2021.13.0
tensorboardX==2.6.2.2
thinc==8.2.5
tim==0.5
tokenizers==0.19.1
torch==2.4.0
torchtext==0.3.1
torchvision==0.19.0
tqdm==4.66.4
transformers==4.43.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
websockets==12.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
